+++ src/Arguments/KeyVault/BaseKeyVaultArguments.cs
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace AzureMcp.Arguments.KeyVault;
+
+public class BaseKeyVaultArguments : SubscriptionArguments
+{
+    public string? VaultName { get; set; }
+}