+++ src/Arguments/KeyVault/Key/KeyGetArguments.cs
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace AzureMcp.Arguments.KeyVault.Key;
+
+public class KeyGetArguments : BaseKeyVaultArguments
+{
+    public string? KeyName { get; set; }
+}