+++ src/Models/Argument/ArgumentDefinitions.cs
@@ -410,4 +410,26 @@ public static class ArgumentDefinitions
             );
         }
     }
+
+    public static class KeyVault
+    {
+        public const string VaultNameParam = "vault";
+        public const string KeyNameParam = "key";
+        public const string KeyTypeParam = "key-type";
+
+        public static readonly ArgumentDefinition<string> VaultName = new(
+            VaultNameParam,
+            "The name of the Key Vault.",
+            required: true);
+
+        public static readonly ArgumentDefinition<string> KeyName = new(
+            KeyNameParam,
+            "The name of the key to retrieve/modify from the Key Vault.",
+            required: true);
+
+        public static readonly ArgumentDefinition<string> KeyType = new(
+            KeyTypeParam,
+            "The type of key to create (RSA, EC).",
+            required: true);
+    }
 }