+++ src/Program.cs
@@ -9,6 +9,7 @@ using AzureMcp.Models;
 using AzureMcp.Models.Command;
 using AzureMcp.Services.Azure.AppConfig;
 using AzureMcp.Services.Azure.Cosmos;
+using AzureMcp.Services.Azure.KeyVault;
 using AzureMcp.Services.Azure.Kusto;
 using AzureMcp.Services.Azure.Monitor;
 using AzureMcp.Services.Azure.Postgres;
@@ -74,6 +75,7 @@ internal class Program
         services.AddSingleton<IAppConfigService, AppConfigService>();
         services.AddSingleton<ISearchService, SearchService>();
         services.AddSingleton<IPostgresService, PostgresService>();
+        services.AddSingleton<IKeyVaultService, KeyVaultService>();
         services.AddSingleton<CommandFactory>();
     }
 }